Reworked the code/some functions act different
Ability to use some commands while bot is not playing (like antispam). Also, will not delete server map when disconnecting and will add server map if it is not there when command is used. Bot will now stay in the voice channel when music is finished and added more accurate detection if user inputs playlist key